+++ fiberplane-models/src/realtime.rs
@@ -1,4 +1,5 @@
 use crate::comments::{Thread, ThreadItem, UserSummary};
+use crate::events::Event;
 use crate::labels::LabelValidationError;
 use crate::notebooks::operations::Operation;
 use base64uuid::Base64Uuid;
@@ -92,6 +93,15 @@ pub enum ServerRealtimeMessage {
     /// Response from a DebugRequest. Contains some useful data regarding the
     /// connection.
     DebugResponse(DebugResponseMessage),
+
+    /// New event was added to the workspace
+    EventAdded(EventAddedMessage),
+
+    /// Event was updated in the workspace
+    EventUpdated(EventUpdatedMessage),
+
+    /// Event was deleted from the workspace
+    EventDeleted(EventDeletedMessage),
 
     /// Notifies a mentioned user of the fact they've been mentioned by someone
     /// else.
@@ -355,6 +365,54 @@ pub struct DebugResponseMessage {
     #[builder(default, setter(into))]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub op_id: Option<String>,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Serialize, TypedBuilder)]
+#[cfg_attr(
+    feature = "fp-bindgen",
+    derive(Serializable),
+    fp(rust_module = "fiberplane_models::realtime")
+)]
+#[non_exhaustive]
+#[serde(rename_all = "camelCase")]
+pub struct EventAddedMessage {
+    /// ID of workspace in which the event was added.
+    pub workspace_id: Base64Uuid,
+
+    /// The event that was added.
+    pub event: Event,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Serialize, TypedBuilder)]
+#[cfg_attr(
+    feature = "fp-bindgen",
+    derive(Serializable),
+    fp(rust_module = "fiberplane_models::realtime")
+)]
+#[non_exhaustive]
+#[serde(rename_all = "camelCase")]
+pub struct EventUpdatedMessage {
+    /// ID of workspace in which the event was updated.
+    pub workspace_id: Base64Uuid,
+
+    /// The event that was updated.
+    pub event: Event,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Eq, Serialize, TypedBuilder)]
+#[cfg_attr(
+    feature = "fp-bindgen",
+    derive(Serializable),
+    fp(rust_module = "fiberplane_models::realtime")
+)]
+#[non_exhaustive]
+#[serde(rename_all = "camelCase")]
+pub struct EventDeletedMessage {
+    /// ID of workspace in which the event was deleted.
+    pub workspace_id: Base64Uuid,
+
+    /// ID of the event that was deleted.
+    pub event_id: String,
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Eq, Serialize, TypedBuilder)]